Move tests tro their own folder
